samplebox-detection/gpuserver/requirements.txt
tensorflow
google-api-python-client
google-cloud-storage
opencv-python
numpy
Flask

